+++ src/io/segment/android/integrations/CountlyIntegration.java
@@ -3,7 +3,7 @@ package io.segment.android.integrations;
 import io.segment.android.errors.InvalidSettingsException;
 import io.segment.android.integration.SimpleIntegration;
 import io.segment.android.models.EasyJSONObject;
-import io.segment.android.models.EventProperties;
+import io.segment.android.models.Props;
 import io.segment.android.models.Screen;
 import io.segment.android.models.Track;
 
@@ -72,7 +72,7 @@ public class CountlyIntegration extends SimpleIntegration {
 	@Override
 	public void track(Track track) {
 		String event = track.getEvent();
-		EventProperties properties = track.getProperties();
+		Props properties = track.getProperties();
 		
 		Map<String, String> segmentation = new HashMap<String, String>();
 		int count = 0;