+++ src/io/segment/android/integrations/FlurryIntegration.java
@@ -4,7 +4,7 @@ import io.segment.android.Logger;
 import io.segment.android.errors.InvalidSettingsException;
 import io.segment.android.integration.SimpleIntegration;
 import io.segment.android.models.EasyJSONObject;
-import io.segment.android.models.EventProperties;
+import io.segment.android.models.Props;
 import io.segment.android.models.Identify;
 import io.segment.android.models.Screen;
 import io.segment.android.models.Track;
@@ -118,12 +118,12 @@ public class FlurryIntegration extends SimpleIntegration {
 	@Override
 	public void track(Track track) {
 		String event = track.getEvent();
-		EventProperties properties = track.getProperties();
+		Props properties = track.getProperties();
 		
 		FlurryAgent.logEvent(event, toMap(properties));
 	}
 
-	private Map<String, String> toMap(EventProperties properties) {
+	private Map<String, String> toMap(Props properties) {
 		Map<String, String> map = new HashMap<String, String>();
 		
 		if (properties != null) {