+++ src/io/segment/android/models/Props.java
@@ -1,0 +1,19 @@
+package io.segment.android.models;
+
+import org.json.JSONObject;
+
+public class Props extends EasyJSONObject {
+
+	public Props() {
+		super();
+	}
+
+	public Props(JSONObject json) {
+		super(json);
+	}
+	
+	public Props(Object... kvs) {
+		super(kvs);
+	}
+	
+}