+++ src/io/segment/android/models/Track.java
@@ -17,7 +17,7 @@ public class Track extends BasePayload {
 	
 	public Track(String userId, 
 				 String event, 
-				 EventProperties properties, 
+				 Props properties, 
 				 Calendar timestamp,
 				 Context context) {
 
@@ -37,13 +37,13 @@ public class Track extends BasePayload {
 		this.put(EVENT_KEY, event);
 	}
 
-	public EventProperties getProperties() {
+	public Props getProperties() {
 		JSONObject object = getObject(PROPERTIES_KEY);
 		if (object == null) return null;
-		else return new EventProperties(object);
+		else return new Props(object);
 	}
 
-	public void setProperties(EventProperties properties) {
+	public void setProperties(Props properties) {
 		this.put(PROPERTIES_KEY, properties);
 	}
 }